requirements.txt
numpy~=1.24.1
matplotlib~=3.7.5
pillow~=10.2.0
torch~=2.3.1+cu121
clip~=1.0
torchvision~=0.18.1+cu121
tqdm~=4.66.4
sacrebleu~=2.4.2
transformers~=4.41.2